Migrate to scala 3.3.1
Build tool was switched to sbt, as in maven supporting 2 different scala
versions is a bit difficult. Post-8 java versions should also work now.
The api should be identical to the scala 2.11 version.
