--- pom.xml
@@ -1,191 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-  <packaging>jar</packaging>
-  <version>1.0.0-SNAPSHOT</version>
-
-  <groupId>com.github.aimmoth.scalajs.compiler.servlet</groupId>
-  <artifactId>scalajs-compiler-servlet</artifactId>
-
-  <properties>
-    <!-- These are set in Config.scala too! -->
-    <scala.version>2.11.12</scala.version> <!-- https://mvnrepository.com/artifact/org.scala-lang/scala-library -->
-    <scalajs.version>0.6.33</scalajs.version> <!-- https://mvnrepository.com/artifact/org.scala-js/scalajs-compiler -->
-    <junit.version>4.13.2</junit.version> <!-- https://mvnrepository.com/artifact/junit/junit -->
-
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <maven.compiler.showDeprecation>true</maven.compiler.showDeprecation>
-  </properties>
-
-  <description>Compiles list of Strings with ScalaJS code to JavaScript. This version uses servlet context to read files on classpath.</description>
-
-  <developers>
-    <developer>
-      <name>Carl Emmoth</name>
-      <url>https://github.com/aimmoth</url>
-    </developer>
-  </developers>
-
-  <prerequisites>
-    <maven>3.1.0</maven>
-  </prerequisites>
-
-  <dependencies>
-    <!-- To add ServletContext -->
-    <dependency>
-      <groupId>javax.servlet</groupId>
-      <artifactId>javax.servlet-api</artifactId>
-      <version>3.1.0</version>
-      <scope>provided</scope>
-    </dependency>
-
-    <!-- Scala -->
-    <dependency>
-      <groupId>org.scala-lang</groupId>
-      <artifactId>scala-library</artifactId>
-      <version>${scala.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.scala-lang</groupId>
-      <artifactId>scala-compiler</artifactId>
-      <version>${scala.version}</version>
-      <scope>provided</scope>
-    </dependency>
-
-    <!-- Scala JS -->
-    <dependency>
-      <groupId>org.scala-js</groupId>
-      <artifactId>scalajs-compiler_2.11.12</artifactId>
-      <version>${scalajs.version}</version>
-    </dependency>
-    <dependency>
-      <!-- https://mvnrepository.com/artifact/org.scala-js/scalajs-library_2.11/1.8.0 -->
-      <groupId>org.scala-js</groupId>
-      <artifactId>scalajs-library_2.11</artifactId>
-      <version>${scalajs.version}</version>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <!-- https://mvnrepository.com/artifact/org.scala-js/scalajs-tools_sjs1.0.0-M2_2.11/1.0.0-M2 -->
-      <groupId>org.scala-js</groupId>
-      <artifactId>scalajs-tools_2.11</artifactId>
-      <version>${scalajs.version}</version>
-    </dependency>
-
-    <!-- Logging -->
-<!--     <dependency> -->
-<!--       <groupId>org.slf4j</groupId> -->
-<!--       <artifactId>slf4j-api</artifactId> -->
-<!--       <version>1.7.21</version> -->
-<!--     </dependency> -->
-<!--     <dependency> -->
-<!--       <groupId>org.slf4j</groupId> -->
-<!--       <artifactId>slf4j-simple</artifactId> -->
-<!--       <version>1.7.21</version> -->
-<!--     </dependency> -->
-
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>${junit.version}</version>
-      <scope>test</scope>
-    </dependency>
-
-  </dependencies>
-
-  <build>
-    <plugins>
-
-      <plugin>
-        <groupId>org.codehaus.mojo</groupId>
-        <artifactId>build-helper-maven-plugin</artifactId>
-        <version>3.2.0</version>
-        <executions>
-          <execution>
-            <id>add-source</id>
-            <phase>generate-sources</phase>
-            <goals>
-              <goal>add-source</goal>
-            </goals>
-            <configuration>
-              <sources>
-                <source>src/main/scala</source>
-              </sources>
-            </configuration>
-          </execution>
-          <execution>
-            <id>add-test-source</id>
-            <phase>generate-sources</phase>
-            <goals>
-              <goal>add-test-source</goal>
-            </goals>
-            <configuration>
-              <sources>
-                <source>src/test/scala</source>
-              </sources>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-
-      <!-- https://davidb.github.io/scala-maven-plugin/plugin-info.html -->
-      <plugin>
-        <groupId>net.alchim31.maven</groupId>
-        <artifactId>scala-maven-plugin</artifactId>
-        <version>4.5.6</version>
-        <executions>
-          <execution>
-            <goals>
-              <goal>compile</goal>
-              <goal>testCompile</goal>
-            </goals>
-          </execution>
-        </executions>
-        <configuration>
-          <sourceDir>src/main/scala</sourceDir>
-          <testSourceDir>src/test/scala</testSourceDir>
-          <scalaVersion>${scala.version}</scalaVersion>
-        </configuration>
-      </plugin>
-
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <version>3.3</version>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
-        </configuration>
-      </plugin>
-
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-jar-plugin</artifactId>
-        <version>3.2.0</version>
-      </plugin>
-    </plugins>
-  </build>
-
-  <reporting>
-    <plugins>
-
-      <plugin>
-        <groupId>org.codehaus.mojo</groupId>
-        <artifactId>versions-maven-plugin</artifactId>
-        <version>2.8.1</version>
-        <reportSets>
-          <reportSet>
-            <reports>
-              <report>dependency-updates-report</report>
-              <report>plugin-updates-report</report>
-              <report>property-updates-report</report>
-            </reports>
-          </reportSet>
-        </reportSets>
-      </plugin>
-
-    </plugins>
-  </reporting>
-</project>
