+++ src/main/scala-3/com/github/aimmoth/scalajs/compiler/servlet/Config.scala
@@ -1,0 +1,13 @@
+package com.aimmoth.scalajs.compiler.servlet
+
+object Config {
+
+  val environments: Map[String, List[String]] = Map[String, List[String]]("default" -> List("dom", "scalatags", "async"))
+
+  val scala2LibraryVersion = "2.13.10"
+  val scalaVersion = "3.3.1"
+  val scalaMainVersion = scalaVersion.split('.').take(2).mkString(".")
+  val scalaJSVersion = "1.12.0"
+  val scalaJSMainVersion = scalaJSVersion.split('.').take(2).mkString(".")
+}
+
